implemented blog-digest querrying
