--- backend-gin/cmd/backend-gin/main.go
@@ -1,134 +1,0 @@
-package main
-
-import (
-	"database/sql"
-	"fmt"
-	"log"
-	"net/http"
-	"os"
-
-	"github.com/gin-gonic/gin"
-	"github.com/joho/godotenv"
-	_ "github.com/lib/pq"
-)
-
-// User represents a user in the system.
-type User struct {
-	ID        int    `json:"id"`
-	Username  string `json:"username"`
-	Email     string `json:"email"`
-	CreatedAt string `json:"created_at"`
-}
-
-var db *sql.DB
-
-// InitializeDatabase sets up the database connection.
-func InitializeDatabase() {
-	// Load environment variables from .env file
-	err := godotenv.Load()
-	if err != nil {
-		log.Println("No .env file found. Using environment variables.")
-	}
-
-	dbHost := getEnv("DB_HOST", "localhost")
-	dbPort := getEnv("DB_PORT", "5432")
-	dbUser := getEnv("DB_USER", "postgres")
-	dbPassword := getEnv("DB_PASSWORD", "")
-	dbName := getEnv("DB_NAME", "postgres")
-	dbSSLMode := getEnv("DB_SSLMODE", "disable")
-
-	psqlInfo := fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=%s",
-		dbHost, dbPort, dbUser, dbPassword, dbName, dbSSLMode)
-
-	db, err = sql.Open("postgres", psqlInfo)
-	if err != nil {
-		log.Fatalf("Error opening database: %v", err)
-	}
-
-	// Verify the connection
-	err = db.Ping()
-	if err != nil {
-		log.Fatalf("Error connecting to the database: %v", err)
-	}
-
-	log.Println("Successfully connected to the database!")
-}
-
-// getEnv retrieves environment variables or returns a default value.
-func getEnv(key, defaultVal string) string {
-	if value, exists := os.LookupEnv(key); exists {
-		return value
-	}
-	return defaultVal
-}
-
-// getUsers handles GET /users requests.
-func getUsers(c *gin.Context) {
-	rows, err := db.Query("SELECT id, username, email, created_at FROM users")
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-	defer rows.Close()
-
-	users := []User{}
-	for rows.Next() {
-		var user User
-		err := rows.Scan(&user.ID, &user.Username, &user.Email, &user.CreatedAt)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			return
-		}
-		users = append(users, user)
-	}
-
-	c.JSON(http.StatusOK, users)
-}
-
-// createUser handles POST /users requests.
-func createUser(c *gin.Context) {
-	var newUser struct {
-		Username string `json:"username" binding:"required"`
-		Email    string `json:"email" binding:"required,email"`
-	}
-
-	if err := c.ShouldBindJSON(&newUser); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	var user User
-	err := db.QueryRow(
-		"INSERT INTO users (username, email) VALUES ($1, $2) RETURNING id, created_at",
-		newUser.Username, newUser.Email,
-	).Scan(&user.ID, &user.CreatedAt)
-	if err != nil {
-		// Check for unique constraint violation
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	user.Username = newUser.Username
-	user.Email = newUser.Email
-
-	c.JSON(http.StatusCreated, user)
-}
-
-func main() {
-	// Initialize the database connection
-	InitializeDatabase()
-	defer db.Close()
-
-	// Create a new Gin router
-	router := gin.Default()
-
-	// Define routes
-	router.GET("/users", getUsers)
-	router.POST("/users", createUser)
-
-	// Start the server on the specified port
-	port := getEnv("PORT", "8080")
-	if err := router.Run(":" + port); err != nil {
-		log.Fatalf("Failed to run server: %v", err)
-	}
-}
